Fixed errors that prevented main from building.
When seting CMAKE_CXX_FLAGS variable, the replacing value did not
have quotes around it. This was due to the quotes not being escaped
so bash ended up only seeing the string inside of the quotes. Also,
corrected the placement of main.cpp from ./main.cpp to ./src/main.cpp